cpu: Dispatch PCEvents to the CheckerCPU SimpleThread
Moving PCEvents scheduling from the system to the Thread [1]
requires us to forward PCEvents to the CheckerCPU thread.

We will otherwise encounter a divergence with the checker when trying to
emulate a SkipFunction in the host (e.g. udelay on Arm).
While the original thread will correctly emulate it and jump to the next
instruction in the binary), the Checker's thread, with no scheduled
PCEvent, will jump straight into the function.

This is fixing realview64-o3-checker.py regression

JIRA: https://gem5.atlassian.net/browse/GEM5-364

[1]: https://gem5-review.googlesource.com/c/public/gem5/+/22106

Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>
Change-Id: I8543535eac3adc366e976b1c0999aafaeca6b141
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/46625
Reviewed-by: Richard Cooper <richard.cooper@arm.com>
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: kokoro <noreply+kokoro@google.com>